Auth flow refactor
