+++ services/textile-hub-auth/package.json
@@ -19,13 +19,14 @@
   "dependencies": {
     "@packages/apitools": "^0.0.1",
     "@packages/models": "^0.0.1",
-    "@textile/hub": "^0.4.1",
+    "@textile/hub": "^0.6.1",
     "aws-lambda": "^1.0.6",
     "aws-sdk": "^2.707.0",
     "emittery": "^0.7.0",
     "ipfs-http-client": "^46.0.0",
     "isomorphic-ws": "^4.0.1",
     "jsonwebtoken": "^8.5.1",
+    "rxjs": "^6.6.2",
     "uuid": "^8.3.0",
     "ws": "^7.3.0"
   },
