+++ services/textile-hub-auth/src/ws/default.ts
@@ -1,0 +1,7 @@
+import { APIGatewayProxyEvent, APIGatewayProxyResult } from 'aws-lambda';
+
+export const handler = async function(
+  event: APIGatewayProxyEvent
+): Promise<APIGatewayProxyResult> {
+  return { statusCode: 200, body: '' };
+}
